OldDemo/scripts/requirements.txt
Flask
flask-cors
langchain
langchain-google-genai
langchain-community
google-generativeai
faiss-cpu
python-dotenv